rag-chatbot-backend/pyproject.toml
[tool.poetry]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "RAG Chatbot Backend"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pydantic = "^2.5.2"
qdrant-client = "^1.6.0"
google-generativeai = "^0.3.1"
psycopg2-binary = "^2.9.9"
tenacity = "^8.2.3"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
unstructured = {extras = ["md"], version = "^0.11.2"}
langchain-text-splitters = "^0.0.1"
beautifulsoup4 = "^4.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
